pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "behavior-lab"
version = "0.1.0"
description = "Modular skeleton-based behavior recognition platform for humans and animals"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "Joon"},
]

dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "scikit-learn>=1.2",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "tqdm>=4.62",
    "einops>=0.6",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
torch = ["torch>=2.0", "torchvision>=0.15"]
dlc = ["deeplabcut[modelzoo]>=3.0.0rc1"]
web = ["fastapi>=0.100", "uvicorn>=0.23", "python-multipart>=0.0.6"]
viz = ["matplotlib>=3.6", "seaborn>=0.12", "imageio>=2.25", "imageio-ffmpeg>=0.4"]
clustering = ["umap-learn>=0.5"]
all = ["behavior-lab[torch,viz,clustering]"]
dev = ["behavior-lab[all]", "pytest>=7.0", "pytest-cov>=4.0", "ruff>=0.1", "mypy>=1.0"]

[project.scripts]
behavior-lab-train = "scripts.train:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
